pyproject.toml
[tool.poetry]
name = "tracy"
version = "0.1.0"
description = "The tracing dice tray"
authors = ["Lucia Prado <prado.lucia1@gmail.com>"]
readme = "README.md"
packages = [
    {include = "tracy", from="src"},
    {include = "tools", from="src"}
]


[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.7.0.72"
apriltag = "^0.0.16"
pyzmq = "^25.0.0"
ultralytics = "^8.0.49"
pygame = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
sim = 'tools.tracy_sim:main'
tracy = 'tracy.main:main'
